agent-service/agent/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-anthropic
langchain-mcp-adapters
langchain-core
fastmcp
httpx
python-dotenv
pydantic
